Fixes .travis.yml and tests that would fail on travis-ci
Travis-CI.org would not pass because there was no databse.yml in the
repo. This commit adds 2 before scripts to copy the database.example.yml
to database.yml and run the migrations. It also includes setting up
notifications for the IRC channel.

The tests for Presentation.past_or_by_month was failing on travis due to
timezone issues. I changed the current date for that test so that the
timezone issues would not affect the test suite.